+++ angular/src/app/core/navigation/navigation.component.scss
@@ -1,0 +1,3 @@
+.navbar-brand {
+  color: white !important;
+}
