+++ angular/src/app/customer/customer-info/customer-info.component.html
@@ -1,0 +1,50 @@
+<h5 class="card-title">Kundendetails</h5>
+<h6 class="card-subtitle mb-2 text-muted">Kundendaten</h6>
+<div class="card-text">
+  <table class="table table-sm table-borderless">
+    <tbody>
+      <tr>
+        <td>Kunde</td>
+        <td>
+          {{ customer?.salutation }} {{ customer?.name }}
+          {{ customer?.surname }}
+        </td>
+      </tr>
+      <tr>
+        <td>Nationalität</td>
+        <td>{{ customer?.nationality }}</td>
+      </tr>
+      <tr>
+        <td>Adresse</td>
+        <td>
+          {{ customer?.address.postalcode }}
+          {{ customer?.address.city }},
+          {{ customer?.address.street }}
+        </td>
+      </tr>
+    </tbody>
+  </table>
+</div>
+<h6 class="card-subtitle mb-2 text-muted">Kontakt</h6>
+<div class="card-text">
+  <table class="table table-sm table-borderless">
+    <tbody>
+      <tr>
+        <td>E-Mail</td>
+        <td>{{ customer?.email }}</td>
+      </tr>
+      <tr>
+        <td>Telefon</td>
+        <td>{{ customer?.phone }}</td>
+      </tr>
+      <tr>
+        <td>Kontakt per</td>
+        <td>
+          <div class="td-badge" *ngFor="let form of customer?.contactform">
+            <span class="badge badge-secondary">{{ form }}</span>
+          </div>
+        </td>
+      </tr>
+    </tbody>
+  </table>
+</div>
