+++ angular/src/app/customer/customer-info/customer-info.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit, Input } from '@angular/core';
+import { Customer } from '../../models/customer';
+
+@Component({
+  selector: 'app-customer-info',
+  templateUrl: './customer-info.component.html',
+  styleUrls: ['./customer-info.component.scss']
+})
+export class CustomerInfoComponent implements OnInit {
+  @Input() customer: Customer;
+  @Input() avatarUrl: string;
+
+  constructor() { }
+
+  ngOnInit() { }
+
+}
