+++ angular/src/app/login/login.component.ts
@@ -36,7 +36,7 @@ export class LoginComponent implements OnInit {
     // If no redirect has been set, use the default
     const redirect = this.authService.redirectUrl
       ? this.router.parseUrl(this.authService.redirectUrl)
-      : '/management';
+      : '/';
 
     // Redirect the user
     this.router.navigateByUrl(redirect);
