+++ angular/src/app/login/login.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { LoginComponent } from './login.component';
+import { RouterModule } from '@angular/router';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    RouterModule
+  ],
+  declarations: [
+    LoginComponent
+  ],
+  exports: [
+    LoginComponent
+  ]
+})
+export class LoginModule { }
