+++ angular/src/app/management/management.module.ts
@@ -4,12 +4,22 @@ import { RouterModule } from '@angular/router';
 import { SharedModule } from '../shared';
 
 import { ManagementComponent } from './management.component';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { NgxSpinnerModule } from 'ngx-spinner';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { NgbCollapseModule, NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 @NgModule({
   imports: [
     CommonModule,
     RouterModule,
-    SharedModule
+    SharedModule,
+    FormsModule,
+    ReactiveFormsModule,
+    NgbModule,
+    NgbCollapseModule,
+    FontAwesomeModule,
+    NgxSpinnerModule
   ],
   declarations: [
     ManagementComponent
